Enable derivation of serde Serialize and Deserialize instances via use_serde flag.
This is needed for loading / saving of state between invocations of a program,
which is probably a common requirement for an algorithm that works on streaming
data.

Unfortunately i could not figure out a way to make the optional package "serde"
depend on the "serde" feature of ordered-float, and i had to use another feature
name instead.